refactor: move constants to constants.js
