+++ fancy-weather/src/js/api/network.js
@@ -1,19 +1,4 @@
-const API_KEYS = {
-  ipgeolocation: 'f91d32ecc60c4fa3a481317e771bc0a5',
-  openweather: '3ca5b2ba64daf09b15dd63cd43add0d3',
-  unsplash: 'hPPxTA7RRG2h0yl4kQgn8LA4BRK288XuhzGBI8akFH4',
-  flickr: '71b96ce362ccd0b4ab0a0239898653aa',
-  opencagedata: 'd86a796b59cc4249bdc9815fab9f7cba',
-  mapbox: 'pk.eyJ1IjoiaXVkaW5hbGVrc2VpIiwiYSI6ImNrYWN3Z2k0dTFrancyem10d2R6dHZzamwifQ.d6tuHi7UozhomnIn3SKAAA',
-  ymaps: 'c110fa3c-57fc-4a15-ae11-02032cd4cdce',
-};
-
-const API_REQUEST = {
-  ipgeolocation: 'https://api.ipgeolocation.io/ipgeo?apiKey=',
-  opencagedata: 'https://api.opencagedata.com/geocode/v1/json?&pretty=1&no_annotations=1',
-  yandexGeocoder: 'https://geocode-maps.yandex.ru/1.x/?format=json&kind=locality',
-  unsplash: 'https://api.unsplash.com/photos/random?orientation=landscape&per_page=1',
-};
+import { API_KEYS, API_REQUEST } from '../constants/constants';
 
 const getDataFromApi = async (url) => {
   try {
