+++ fancy-weather/src/js/dom/currentWeather.js
@@ -1,61 +1,5 @@
+import { PAGE_ELEMENT, WEATHER_CONDITION_LIST, UNITS } from '../constants/constants';
 import { getTemperatureValue, roundWindSpeed } from '../utils/round';
-
-const PAGE_ELEMENT = {
-  currentWeather: 'weather__current',
-  currentTemperature: 'weather__current__temp',
-  weatherIcon: 'weather__current__icon',
-  currentWeatherCondition: 'weather__current__condition',
-  currentWeatherList: 'weather__current__list',
-  forecastWeather: 'weather__forecast',
-};
-
-const WEATHER_CONDITION_LIST = {
-  en: {
-    description: '',
-    feels_like: 'feels like: ',
-    wind_speed: 'wind speed: ',
-    humidity: 'humidity: ',
-    uvi: 'UV-index: ',
-  },
-  ru: {
-    description: '',
-    feels_like: 'ощущается как: ',
-    wind_speed: 'скорость ветра: ',
-    humidity: 'влажность: ',
-    uvi: 'УФ-индекс: ',
-  },
-  be: {
-    description: '',
-    feels_like: 'адчуваецца як: ',
-    wind_speed: 'хуткасць ветру: ',
-    humidity: 'вільготнасць: ',
-    uvi: 'УФ-індэкс: ',
-  },
-};
-
-const UNITS = {
-  en: {
-    description: '',
-    feels_like: '°',
-    wind_speed: ' m/s',
-    humidity: ' %',
-    uvi: '',
-  },
-  ru: {
-    description: '',
-    feels_like: '°',
-    wind_speed: ' м/с',
-    humidity: ' %',
-    uvi: '',
-  },
-  be: {
-    description: '',
-    feels_like: '°',
-    wind_speed: ' м/с',
-    humidity: ' %',
-    uvi: '',
-  },
-};
 
 const createCurrentTemperature = (temp, units) => {
   const temperature = document.createElement('p');
