+++ fancy-weather/src/js/utils/messageForUser.js
@@ -1,6 +1,4 @@
-const PAGE_ELEMENT = {
-  message: 'search__message',
-};
+import PAGE_ELEMENT from '../constants/constants';
 
 const setUserMessage = (text) => {
   const messageContainer = document.querySelector(`.${PAGE_ELEMENT.message}`);
